Fix a few bugs introduced with file locking: Only use one instance of lock manager. Also, we need to be careful about locking during partial output removal

+++ src/main/scala/ducttape/exec/PartialOutputMover.scala
@@ -22,21 +22,28 @@ import grizzled.slf4j.Logging
  */
 class PartialOutputMover(dirs: DirectoryArchitect,
                          partial: Set[(String,Realization)],
-                         broken: Set[(String,Realization)]) extends UnpackedDagVisitor with Logging {
+                         broken: Set[(String,Realization)],
+                         locker: LockManager) extends UnpackedDagVisitor with Logging {
   
   override def visit(task: RealTask) {
     
     debug("Considering %s".format(task))
     
-    if (broken( (task.name, task.realization) )) {
-      System.err.println("Removing broken partial output for %s".format(task))
-      val origDir = dirs.assignDir(task)
-      Files.deleteDir(origDir)
-      
-    } else if (partial( (task.name, task.realization) )) {
-      System.err.println("Moving %s to the attic".format(task))
-      val taskEnv = new TaskEnvironment(dirs, task)
-      PartialOutputMover.moveToAttic(taskEnv)
+    val taskEnv = new TaskEnvironment(dirs, task)
+    val gotLock = locker.maybeAcquireLock(taskEnv)
+    
+    if (gotLock) {
+      if (broken( (task.name, task.realization) )) {
+        System.err.println("Removing broken partial output for %s".format(task))
+        val origDir = dirs.assignDir(task)
+        Files.deleteDir(origDir)
+        
+      } else if (partial( (task.name, task.realization) )) {
+        System.err.println("Moving %s to the attic".format(task))
+        PartialOutputMover.moveToAttic(taskEnv)
+      }
+    } else {
+      debug("Couldn't immediately get a lock for %s. We'll let the other process keep using it for now and try again later".format(task))
     }
   }
 }
